+++ src/main/java/com/project/joonggo/domain/ChatJoinVO.java
@@ -9,5 +9,5 @@ import lombok.*;
 @NoArgsConstructor
 public class ChatJoinVO {
     private int roomId;             // 채팅방 id
-    private int userNum;            // 참여자 num. id는 String 임.
+    private long userNum;            // 참여자 num. id는 String 임.
 }
